// src/libs/models/challenge.ts
/**
 * Miata Rpg API
 * This is a simple API
 *
 * OpenAPI spec version: 1.0.0
 * Contact: you@your-company.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Difficulty } from './difficulty';
import { FocusNode } from './focusNode';
import { Material } from './material';
import { Util } from './util';

export interface ChallengeStep {
    description: string,
    status: boolean,
    title: string,
    subTitle: string,
    disabled?: boolean
}

export interface Challenge { 
    name: string;
    completed: boolean;
    completedOn?: string;
    timeToComplete?: string;
    description: string;
    steps: ChallengeStep[];
    focus: FocusNode;
    difficulty: Difficulty;
    materials: Array<Material>;
    utils: Array<Util>;
}